requirements.txt
streamlit==1.35.0
pandas==2.2.2
numpy==1.24.3
matplotlib==3.7.1
scikit-learn==1.2.2
tensorflow-cpu==2.12.0
keras==2.12.0